server turns off finally lol

+++ torchboard/server/torchboard_server.py
@@ -1,15 +1,17 @@
 import flask
 from flask_cors import CORS, cross_origin
 from flask_session import Session
+from werkzeug.serving import make_server
 
-from multiprocessing import Process, Manager
+from threading import Thread, Event
 
 import webbrowser
 
 from typing import Any, List
 
 import os
-from torchboard.server.utils import wipe_dir  
+from torchboard.server.utils import wipe_dir
+
 class TorchBoardServer():
     port:int
     host:str
@@ -23,14 +25,13 @@ class TorchBoardServer():
     def __init__(self, port:int=8080, host:str='127.0.0.1', name:str='TorchBoard', static_path:str='static') -> None:
         self.port = port
         self.host = host
-        
-        self.__mp_manager = Manager()
-        
-        self.listener_state = self.__mp_manager.dict()
-        self.variable_state = self.__mp_manager.dict()
+    
+        self.listener_state = dict()
+        self.variable_state = dict()
         
         self.app = flask.Flask(name)
         self.__flask_process = None
+        self.__stop_event = Event()
         
         self.app.config["SECRET_KEY"] = "duparomana123" #TODO: Change secret key
         self.app.config['SESSION_TYPE'] = 'filesystem'
@@ -50,6 +51,8 @@ class TorchBoardServer():
         self.app.add_url_rule('/get_history','get_history', self.__get_history, methods=['GET'])
         self.app.add_url_rule('/get_variables','get_variables', self.__get_variables, methods=['GET'])
         self.app.add_url_rule('/update_variable','update_variable', self.__update_variable, methods=['PUT'])
+        
+        self.server = make_server(self.host, self.port, self.app)
         
     @cross_origin()
     def __get_changes_session(self) -> flask.Response:
@@ -95,26 +98,35 @@ class TorchBoardServer():
         return flask.jsonify({'status': 'success'}),200
 
     def __run_flask(self) -> None:
-        self.app.run(port=self.port, host=self.host)
+        self.server.serve_forever()
     
-    def run(self) -> None:
+    def start(self) -> None:
         if self.__flask_process:
             return
-        self.__flask_process = Process(target=self.__run_flask, name=self.app.name) #Non blocking
+        self.__flask_process = Thread(target=self.__run_flask)
         self.__flask_process.start()
         webbrowser.open(f'http://{self.host}:{self.port}') #Force open browser to dashboard
     
     def stop(self) -> None:
-        self.__flask_process.terminate()
-        self.__flask_process.join()
-        self.__flask_process = None
+        if not self.__flask_process:
+            return
+        print('Stopping server')
+        
+        
         wipe_dir('flask_session') 
         os.rmdir('flask_session') 
+        
+        self.server.shutdown()
+        print('Server stopped')
+        self.__flask_process.join()
+        self.__flask_process = None
+        
+        #return NotImplementedError('Not implemented yet')
         
     
     def add_listener_variable(self, name:str, value:Any) -> None:
         if name not in self.listener_state:
-            self.listener_state[name] = self.__mp_manager.list()
+            self.listener_state[name] = []
         self.listener_state[name].append(value)
     
     def add_listener_variables(self, variables:dict[str, Any]) -> None:
@@ -144,12 +156,12 @@ class TorchBoardServer():
     
         
 if __name__ == '__main__':
-    import time, sys
+    import time
     server = TorchBoardServer(static_path='../../static')
     server.register_variable('test_int', 2137)
     server.register_variable('test_str', 'Hello')
     server.register_variable('test_float', 0.534)
-    server.run()
+    server.start()
     try:
         i = 1
         while True:
@@ -158,4 +170,4 @@ if __name__ == '__main__':
             time.sleep(1)
     except KeyboardInterrupt:
         server.stop()
-    
+
